Refactor test cases
